simple-chatbot/requirements.txt
openai
streamlit
python-dotenv
transformers
torch
numpy
